requirements.txt
python_telegram_bot==20.1
emoji
requests